% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covHall.R
\name{covHall}
\alias{covHall}
\title{covHall}
\usage{
covHall(data, u, bf = 10, pve = 0.9, spectral_decomp = T, eps = 0.01,
  nu = 1, mu.fit = NULL)
}
\arguments{
\item{mu.fit}{}
}
\description{
Internal function used to compute the latent covariance
}
\author{
Jan Gertheiss \email{jan.gertheiss@agr.uni-goettingen.de}
}

