% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deriv.inv.logit.R
\name{deriv.inv.logit}
\alias{deriv.inv.logit}
\title{deriv.inv.logit}
\usage{
\method{deriv}{inv.logit}(x)
}
\arguments{
\item{x}{Value at which derivative is computed}
}
\description{
Internal function used to compute the derivative of the logit function
}
\author{
Jan Gertheiss \email{jan.gertheiss@agr.uni-goettingen.de}
}

