% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpca-package.R
\docType{package}
\name{gfpca-package}
\alias{gfpca}
\alias{gfpca-package}
\title{Generalized functional principal components analysis}
\description{
A package for estimating functional principal components models
for generalized responses. Current implementations support binary responses.
Several estimation techniques are available; see \code{\link{gfpca_Mar}},
\code{\link{gfpca_TwoStep}}, and \code{\link{gfpca_Bayes}}
}

