% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfpca_Bayes.R
\name{gfpcaBayes}
\alias{gfpcaBayes}
\title{gfpca_Bayes}
\usage{
gfpcaBayes(data, npc = 3, grid = NULL, nbasis = 10, iter = 1000,
  warmup = 400)
}
\arguments{
\item{nbasis}{}
}
\description{
Implements a Bayesian approach to generalized functional principal components analysis for 
sparsely observed binary curves
}
\author{
Jan Gertheiss \email{jan.gertheiss@agr.uni-goettingen.de}
}
\references{
Gertheiss, J., Goldsmith, J., and Staicu, A.-M. (2016).
A note on modeling sparse exponential-family functional response curves. 
\emph{Under Review}.
}
\seealso{
\code{\link{gfpca_TwoStep}}, \code{\link{gfpca_Bayes}}.
}

