% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stanmodel.R
\name{make_stanmodel}
\alias{make_stanmodel}
\title{Make Stan Model from .Stan file}
\usage{
make_stanmodel(mod_name = "gfpca")
}
\arguments{
\item{mod_name}{Name of the .stan file}
}
\value{
Object of class \code{stanmodel}
}
\description{
Simple wrapper to make stanmodel
}

