% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_gfpca.R
\docType{data}
\name{model_gfpca}
\alias{model_gfpca}
\title{gfpca model}
\format{An object of class \code{"stanmodel"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(gfpca)
}
\description{
GFPCA model already compiled
}
\examples{
model_gfpca
}
\keyword{datasets}

