% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSc.R
\name{predSc}
\alias{predSc}
\title{predSc}
\usage{
predSc(ev, psi, Yi.obs, mu, gs)
}
\arguments{
\item{ev}{vector of eigenvalues}

\item{psi}{matrix of eigenfunctions}

\item{Yi.obs}{matrix of observed data}

\item{mu}{mean function}

\item{gs}{dispersion parameter (gamma in Hall et al.). This is gm in gfpcaMar. In
the simulation studies this parameter is chosen such the MSE is smallest,
which gives somewhat over-optimistic results for the marginal approach.}
}
\description{
Internal function for prediction of scores using the Hall et al. approach.
}
\author{
Jan Gertheiss \email{jan.gertheiss@agr.uni-goettingen.de} and 
Ana-Maria Staicu \email{astaicu@ncsu.edu}
}

