% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predSc.R
\name{predSc}
\alias{predSc}
\title{predSc}
\usage{
predSc(ev, psi, Yi.obs, mu, gs)
}
\arguments{
\item{gs}{}
}
\description{
Internal function used to compute predicted scores
}
\author{
Jan Gertheiss \email{jan.gertheiss@agr.uni-goettingen.de}
}

